rename files to fix capitalization
